Split out parser tests
