+++ statham/dsl/parser.py
@@ -17,6 +17,7 @@ _KEYWORD_MAPPER: DefaultDict[str, Callable] = defaultdict(
 
 
 def parse(schema: Dict[str, Any]) -> Element:
+    """Parse a JSONSchema dictionary to a DSL Element."""
     if isinstance(schema.get("type"), list):
         return AnyOf(
             *(parse({**schema, "type": type_}) for type_ in schema["type"])
